+++ History.md
@@ -1,3 +1,18 @@
+## v1.6.2
+- Exposed new `IActorRuntime.GetCurrentActorIds()` API that returns the `ActorId` for each active
+  actor managed by the runtime, as well as an `IActorRuntime.GetCurrentActorTypes()` API that
+  returns the `Type` of each active actor managed by the runtime. These APIs are not thread-safe and
+  should only be used for gathering statistics and debugging purposes.
+
+## v1.6.1
+- Exposed new `IActorRuntime.GetActorExecutionStatus(id)` API that enables querying the actor
+  runtime for the current execution status of the actor with the specified id, as well as an
+  `IActorRuntime.GetCurrentActorCount()` API that returns the number of active actors managed by the
+  runtime. These APIs are not thread-safe and should only be used for gathering statistics and
+  debugging purposes.
+- Exposed new `IActorRuntime.OnActorHalted` callback which is triggered when an actor has halted and
+  the runtime has stopped managing it.
+
 ## v1.6.0
 - Exposed new `Operation` API that enables instrumenting, controlling and scheduling custom
 concurrent operations.